+++ src/Fakturownia.php
@@ -97,6 +97,30 @@ class Fakturownia
     }
 
     /**
+     * Get invoice as PDF
+     *
+     * @param int         $id          Invoice ID
+     * @param string|null $printOption Print option
+     *
+     * @return ResponseInterface
+     *
+     * @throws Exception\RequestErrorException
+     */
+    public function getInvoicePdf(int $id, string $printOption = null): ResponseInterface
+    {
+        $url = $this->baseUrl . '/invoices/' . $id . '.pdf';
+        $params = [
+            'api_token' => $this->apiToken,
+        ];
+
+        if (null !== $printOption) {
+            $params['print_option'] = $printOption;
+        }
+
+        return $this->restClient->get($url, $params);
+    }
+
+    /**
      * Create invoice
      *
      * @param array $invoice Invoice data
@@ -952,6 +976,83 @@ class Fakturownia
     public function deleteDepartment(int $id): ResponseInterface
     {
         $url = $this->baseUrl . '/departments/' . $id . '.json';
+        $params = [
+            'api_token' => $this->apiToken,
+        ];
+
+        return $this->restClient->delete($url, $params);
+    }
+
+    /**
+     * Get price lists
+     *
+     * @param array $params Parameters
+     *
+     * @return ResponseInterface
+     *
+     * @throws Exception\RequestErrorException
+     */
+    public function getPriceLists(array $params = []): ResponseInterface
+    {
+        $url = $this->baseUrl . '/price_lists.json';
+        $params['api_token'] = $this->apiToken;
+
+        return $this->restClient->get($url, $params);
+    }
+
+    /**
+     * Create price list
+     *
+     * @param array $priceList Price list data
+     *
+     * @return ResponseInterface
+     *
+     * @throws Exception\RequestErrorException
+     */
+    public function createPriceList(array $priceList): ResponseInterface
+    {
+        $url = $this->baseUrl . '/price_lists.json';
+        $data = [
+            'price_list' => $priceList,
+            'api_token' => $this->apiToken,
+        ];
+
+        return $this->restClient->post($url, $data);
+    }
+
+    /**
+     * Update price list
+     *
+     * @param int   $id        Price list ID
+     * @param array $priceList Price list data
+     *
+     * @return ResponseInterface
+     *
+     * @throws Exception\RequestErrorException
+     */
+    public function updatePriceList(int $id, array $priceList): ResponseInterface
+    {
+        $url = $this->baseUrl . '/price_lists/' . $id . '.json';
+        $data = [
+            'price_list' => $priceList,
+            'api_token' => $this->apiToken,
+        ];
+
+        return $this->restClient->put($url, $data);
+    }
+
+    /**
+     * Delete price list
+     *
+     * @param int $id Price list ID
+     *
+     * @return ResponseInterface
+     *
+     * @throws Exception\RequestErrorException
+     */
+    public function deletePriceList(int $id): ResponseInterface
+    {
+        $url = $this->baseUrl . '/price_lists/' . $id . '.json';
         $params = [
             'api_token' => $this->apiToken,
         ];